+++ src/package/store.rs
@@ -124,9 +124,10 @@ impl PackageStore {
     pub async fn resolve(&self, package: &PackageName) -> miette::Result<Manifest> {
         let manifest = self.locate(package).join(MANIFEST_FILE);
 
-        let manifest = Manifest::try_read_from(manifest)
-            .await?
-            .ok_or(miette!("the package store is corrupted"))?;
+        let manifest = Manifest::try_read_from(&manifest).await?.ok_or(miette!(
+            "the package store is corrupted: `{}` is not present",
+            manifest.display()
+        ))?;
 
         Ok(manifest)
     }
