--- tests/cmd/publish/in
@@ -1,1 +1,0 @@
-../../data/projects/lib
