+++ tests/cmd/publish/lib/in
@@ -1,0 +1,1 @@
+../../../data/projects/lib
