+++ tests/cmd/publish/lib/out
@@ -1,0 +1,1 @@
+../../../data/projects/lib
