+++ tests/cmd/publish/lib/stdout.log
