+++ tests/cmd/publish/local/in/Proto.lock
@@ -1,0 +1,2 @@
+version = 1
+packages = []
