+++ tests/cmd/publish/local/in/proto/local.proto
@@ -1,0 +1,3 @@
+syntax = "proto3";
+
+package local.api;
