+++ tests/cmd/publish/local/in/some-local-lib/Proto.lock
@@ -1,0 +1,2 @@
+version = 1
+packages = []
