+++ tests/cmd/publish/local/in/some-local-lib/Proto.toml
@@ -1,0 +1,6 @@
+[package]
+type = "lib"
+name = "some-local-lib"
+version = "0.1.0"
+
+[dependencies]
