+++ tests/cmd/publish/local/in/some-local-lib/proto/local.proto
@@ -1,0 +1,3 @@
+syntax = "proto3";
+
+package local.lib;
