+++ tests/cmd/publish/local/stderr.log
@@ -1,0 +1,5 @@
+Error:   × failed to publish `my-api` to `https://localhost:54321/fake-uri:my-
+  │ repository`
+  ╰─▶ unable to public my-api to artifactory due to local dependency some-
+      local-lib
+
