+++ tests/cmd/publish/local/stdout.log
@@ -1,0 +1,1 @@
+:: packaged my-api@0.1.0
