--- tests/cmd/publish/out
@@ -1,1 +1,0 @@
-../../data/projects/lib
